+++ intelcamp/algo_main_gru.py
