+++ intelcamp/algo_main_rnn_v2.py
@@ -2,8 +2,8 @@ import os
 import numpy as np
 import pandas as pd
 import json
-from util import prtime, factors, tile, get_exp_dir
-import rnn
+from intelcamp.util import prtime, factors, tile
+from intelcamp import rnn
 
 import torch
 from torch.utils.data.dataloader import DataLoader
@@ -237,22 +237,9 @@ def save_model(model, epoch, n_iter):
     # prtime("RNN model checkpoint saved")
 
 
-def pinball_np(output, target, configs):
-    resid = target - output
-    tau = np.array(configs["qs"])
-    alpha = configs["smoothing_alpha"]
-    log_term = np.zeros_like(resid)
-    log_term[resid < 0] = (np.log(1+np.exp(resid[resid < 0]/alpha)) - (resid[resid < 0]/alpha))
-    log_term[resid >= 0] = np.log(1 + np.exp(-resid[resid >= 0]/alpha))
-    loss = resid * tau + alpha * log_term
-
-    return loss
-
-
 def test_processing(val_df, val_loader, model, seq_dim, input_dim, val_batch_size, transformation_method, configs):
     """
     Process the val set and report error statistics
-
     :param val_df: DataFrame
     :param val_loader: Data Loader
     :param model:
@@ -263,36 +250,30 @@ def test_processing(val_df, val_loader, model, seq_dim, input_dim, val_batch_siz
     :param configs:
     :return:
     """
-
-    # Plug the val set into the model
+    # val_df, val_loader, model, seq_dim, input_dim, val_batch_size, transformation_method
     model.eval()
+    # val_y_at_t = torch.zeros(val_batch_size, seq_dim, 1)
     preds = []
     targets = []
     for i, (feats, values) in enumerate(val_loader):
+        # features = Variable(feats.view(-1, seq_dim, input_dim - 1))
+        # outputs = model(torch.cat((features, val_y_at_t), dim=2))
+        # val_y_at_t = tile(outputs.unsqueeze(2), 1, 5)
         features = Variable(feats.view(-1, seq_dim, input_dim))
         outputs = model(features)
         preds.append(outputs.data.numpy().squeeze())
         targets.append(values.data.numpy())
 
-    # (Normalized Data) Concatenate the predictions and targets for the whole val set
+    # concatenating the preds and targets for the whole epoch (iterating over val_loader once)
     semifinal_preds = np.concatenate(preds).ravel()
     semifinal_targs = np.concatenate(targets).ravel()
-
-    # (Normalized Data) Assign target and output variables
-    target = semifinal_targs
-    output = semifinal_preds
-    resid = target - output
-
-    # Calculate pinball loss (done on normalized data)
-    loss = pinball_np(semifinal_preds, semifinal_targs, configs)
-    pinball_loss = np.mean(loss)
+    mse_loss = np.mean((semifinal_targs - semifinal_preds) ** 2)
 
     # loading the training data stats for de-normalization purpose
     file_loc = file_prefix + '/train_stats.json'
     with open(file_loc, 'r') as f:
         train_stats = json.load(f)
 
-    # Get normalization statistics
     train_max = pd.DataFrame(train_stats['train_max'], index=[1]).iloc[0]
     train_min = pd.DataFrame(train_stats['train_min'], index=[1]).iloc[0]
     train_mean = pd.DataFrame(train_stats['train_mean'], index=[1]).iloc[0]
@@ -303,49 +284,26 @@ def test_processing(val_df, val_loader, model, seq_dim, input_dim, val_batch_siz
         final_preds = ((train_max[configs['target_var']] - train_min[configs['target_var']]) * semifinal_preds) + \
                       train_min[
                           configs['target_var']]
-        final_targs = ((train_max[configs['target_var']] - train_min[configs['target_var']]) * semifinal_targs) + \
-                      train_min[
-                          configs['target_var']]
 
-    # else:
-    #     final_preds = ((semifinal_preds * train_std[configs['target_var']]) + train_mean[configs['target_var']])
+    else:
+        final_preds = ((semifinal_preds * train_std[configs['target_var']]) + train_mean[configs['target_var']])
 
     predictions = pd.DataFrame(final_preds)
-    output = final_preds
-    target = final_targs
-
-    # Calculate different error metrics for the requested quantile
-    # Set "Number of adjustable model parameters" for each type of error statistic
-    p_nmbe = 0
+    predicted = predictions.values.squeeze()
+    measured = val_df[configs['target_var']].values
+    p_nmbe = 0  # Number of "adjustable model parameters"
     p_cvrmse = 1
-    rmse = np.sqrt(np.mean((output - target) ** 2))
-    nmbe = (1 / (np.mean(target))) * (np.sum(target - output)) / (len(target) - p_nmbe)
-    cvrmse = (1 / (np.mean(target))) * np.sqrt(np.sum((target-output)**2) / (len(target) - p_cvrmse))
-    gof = (np.sqrt(2) / 2) * np.sqrt(cvrmse**2 + nmbe**2)
+
+    # Calculate different error metrics
+    rmse = np.sqrt(np.mean((predicted - measured) ** 2))
+    nmbe = (1 / (np.mean(measured))) * (np.sum(measured - predicted)) / (len(measured) - p_nmbe)
+    cvrmse = (1 / (np.mean(measured))) * np.sqrt(np.sum((measured-predicted)**2) / (len(measured) - p_cvrmse))
+    gof = (np.sqrt(2) / 2)  * np.sqrt(cvrmse**2 + nmbe**2)
 
     # Add different error statistics to a dictionary
-    errors = {"pinball_loss": pinball_loss,
-              "rmse": rmse,
-              "nmbe": nmbe,
-              "cvrmse": cvrmse,
-              "gof": gof
-              }
+    errors = {"mse_loss": mse_loss, "rmse": rmse, "nmbe": nmbe, "cvrmse": cvrmse, "gof": gof}
 
     return predictions, errors
-
-
-def quantile_loss(output, target, configs):
-
-    resid = target - output
-    tau = torch.FloatTensor(configs["qs"])
-    alpha = configs["smoothing_alpha"]
-    log_term = torch.zeros_like(resid)
-    log_term[resid < 0] = (torch.log(1+torch.exp(resid[resid < 0]/alpha)) - (resid[resid < 0]/alpha))
-    log_term[resid >= 0] = torch.log(1 + torch.exp(-resid[resid >= 0]/alpha))
-    loss = resid * tau + alpha * log_term
-    loss = torch.mean(loss)
-
-    return loss
 
 
 def process(train_loader, val_loader, val_df, num_epochs, run_train, run_resume, writer, transformation_method,
@@ -422,7 +380,7 @@ def process(train_loader, val_loader, val_df, num_epochs, run_train, run_resume,
         cuda_avail = torch.cuda.is_available()
 
         # Instantiating Loss Class
-        #criterion = nn.MSELoss()
+        criterion = nn.MSELoss()
 
         # Instantiate Optimizer Class
         optimizer = torch.optim.Adam(model.parameters(), lr=configs['lr_config']['base'], weight_decay=weight_decay)
@@ -486,8 +444,7 @@ def process(train_loader, val_loader, val_df, num_epochs, run_train, run_resume,
                 # train_y_at_t_nump = train_y_at_t.detach().numpy()
 
                 # Calculate Loss: softmax --> cross entropy loss
-                #loss = criterion(outputs.squeeze(), target)
-                loss = quantile_loss(outputs.squeeze(), target, configs)
+                loss = criterion(outputs.squeeze(), target)
 
                 train_loss.append(loss.data.item())
                 train_iter.append(n_iter)
@@ -521,18 +478,18 @@ def process(train_loader, val_loader, val_df, num_epochs, run_train, run_resume,
                                                       "Step": time6 - time5}, n_iter)
 
                 # save the model every ___ iterations
-                if n_iter % 5 == 0:
+                if n_iter % 200 == 0:
                     save_model(model, epoch, n_iter)
 
                 # Do a val batch every ___ iterations
-                if n_iter % 5 == 0:
+                if n_iter % 200 == 0:
                     # Evaluate val set
                     predictions, errors = test_processing(val_df, val_loader, model, seq_dim, input_dim,
                                                           val_batch_size, transformation_method, configs)
                     val_iter.append(n_iter)
-                    #val_loss.append(errors['pinball_loss'])
+                    #val_loss.append(errors['mse_loss'])
                     #val_rmse.append(errors['rmse'])
-                    writer.add_scalars("Loss", {"val": errors['pinball_loss']}, n_iter)
+                    writer.add_scalars("Loss", {"val": errors['mse_loss']}, n_iter)
 
                     # Add matplotlib plot to TensorBoard to compare actual val set vs predicted
                     fig1, ax1 = plt.subplots(figsize=(20, 5))
@@ -554,9 +511,9 @@ def process(train_loader, val_loader, val_df, num_epochs, run_train, run_resume,
                     ax2.axis('equal')
                     writer.add_figure('Parody', fig2, n_iter)
 
-                    print('Epoch: {} Iteration: {}. Train_loss: {}. val_loss: {}, LR: {}'.format(epoch, n_iter,
+                    print('Epoch: {} Iteration: {}. Train_MSE: {}. val_MSE: {}, LR: {}'.format(epoch, n_iter,
                                                                                                 loss.data.item(),
-                                                                                                errors['pinball_loss'],
+                                                                                                errors['mse_loss'],
                                                                                                 optimizer.param_groups[
                                                                                                     0]['lr']))
 
@@ -579,10 +536,10 @@ def process(train_loader, val_loader, val_df, num_epochs, run_train, run_resume,
 
         predictions, errors = test_processing(val_df, val_loader, model, seq_dim, input_dim,
                                               val_batch_size, transformation_method, configs)
-        #val_loss.append(errors['pinball_loss'])
+        #val_loss.append(errors['mse_loss'])
         #val_rmse.append(errors['rmse'])
-        writer.add_scalars("Loss", {"val": errors['pinball_loss']})
-        prtime('val_loss: {}'.format(errors['pinball_loss']))
+        writer.add_scalars("Loss", {"val": errors['mse_loss']})
+        prtime('val_MSE: {}'.format(errors['mse_loss']))
 
         # Save the final predictions and error statistics to a file
         predictions.to_csv(file_prefix + '/predictions.csv', index=False)
@@ -603,16 +560,13 @@ def process(train_loader, val_loader, val_df, num_epochs, run_train, run_resume,
 
 def eval_trained_model(file_prefix, train_data, train_batch_size, configs):
     """
-    Pass the entire training set through the trained model and get the predictions.
-    Compute the residual and save to a DataFrame.
-
+    Pass the entire training set through the trained model and get the predictions. Compute the residual and save to a DataFrame.
     :param file_prefix:
     :param train_data:
     :param train_batch_size:
     :param configs:
     :return:
     """
-
     # Evaluate the training model
     torch_model = torch.load(file_prefix + '/torch_model')
     model = torch_model['torch_model']
@@ -651,7 +605,7 @@ def eval_trained_model(file_prefix, train_data, train_batch_size, configs):
     mask = pd.read_hdf(mask_file, key='df')
     msk = mask["msk"]
 
-    # # Adjust the datetime index so it is in line with the EC data
+    # Adjust the datetime index so it is in line with the EC data
     # target_index = data_time_index[msk] + pd.DateOffset(
     #     minutes=(configs["EC_future_gap"] * configs["resample_freq"]))
     # processed_data = pd.DataFrame(index=target_index)
@@ -713,8 +667,8 @@ def main(train_df, val_df, configs):
     arch_type = configs['arch_type']
 
     global file_prefix
-    file_prefix = get_exp_dir(configs)
-    pathlib.Path(file_prefix).mkdir(parents=True, exist_ok=True)
+    file_prefix = pathlib.Path(configs["exp_dir"])
+    file_prefix.mkdir(parents=True, exist_ok=True)
 
     writer_path = file_prefix
     writer = SummaryWriter(writer_path)
