+++ intelcamp/algo_main_rnn_v5.py
@@ -3,9 +3,9 @@ import os
 import numpy as np
 import pandas as pd
 import json
-from util import prtime, factors, tile, get_exp_dir
-import rnn
-import lstm
+from intelcamp.util import prtime, factors, tile
+from intelcamp import rnn
+from intelcamp import lstm
 
 import torch
 from torch.utils.data.dataloader import DataLoader
@@ -22,7 +22,7 @@ import csv
 import pathlib
 import psutil
 from psutil import virtual_memory
-import buildings_processing as bp
+import intelcamp.buildings_processing as bp
 import logging
 import matplotlib.dates as mdates
 
@@ -761,13 +761,14 @@ def run_validation(val_loader, val_df, writer, transformation_method, configs, v
         json.dump(errors, fp, indent=1)
 
     # If a training history csv file does not exist, make one
-    if not pathlib.Path("Testing_history.csv").exists():
-        with open(r'Testing_history.csv', 'a') as f:
+    train_history_path = file_prefix / "Testing_history.csv"
+    if not train_history_path.exists():
+        with open(train_history_path, 'a') as f:
             writer = csv.writer(f, lineterminator='\n')
             writer.writerow(["File Path", "RMSE", "CV(RMSE)", "NMBE", "GOF", "QS", "ACE", "IS"])
 
     # Save the errors statistics to a central results csv once everything is done
-    with open(r'Testing_history.csv', 'a') as f:
+    with open(train_history_path, 'a') as f:
         writer = csv.writer(f, lineterminator='\n')
         writer.writerow([file_prefix,
                             errors["rmse"],
@@ -783,7 +784,8 @@ def run_validation(val_loader, val_df, writer, transformation_method, configs, v
 
 
 def run_prediction(val_loader, val_df, writer, transformation_method, configs, val_batch_size, seq_dim):
-    torch_model = torch.load(os.path.join(file_prefix, 'torch_model'))
+    torch_model_path = pathlib.Path(configs["exp_dir"]) / 'torch_model'
+    torch_model = torch.load(torch_model_path)
     model = torch_model['torch_model']
     model.eval()
 
@@ -798,8 +800,8 @@ def run_prediction(val_loader, val_df, writer, transformation_method, configs, v
             outputs = model(features)
             preds.append(outputs.cpu().numpy())
 
-        file_loc = os.path.join(file_prefix, "train_stats.json")
-        with open(file_loc, 'r') as f:
+        file_path = pathlib.Path(configs["exp_dir"]) / "train_stats.json"
+        with open(file_path, 'r') as f:
             train_stats = json.load(f)
 
         semifinal_preds = np.concatenate(preds)
@@ -1213,7 +1215,7 @@ def predict(data, file_prefix):
     train_feat_tensor = torch.from_numpy(data).type(torch.FloatTensor)
 
     # Load model
-    torch_model = torch.load(os.path.join(file_prefix, 'torch_model'))
+    torch_model = pathlib.Path(configs["exp_dir"]) / 'torch_model'
     model = torch_model['torch_model']
 
     # Evaluate model
@@ -1269,7 +1271,7 @@ def main(train_df, val_df, configs):
 
     # Make results directory
     global file_prefix
-    file_prefix = get_exp_dir(configs)
+    file_prefix = pathlib.Path(configs["exp_dir"])
     file_prefix.mkdir(parents=True, exist_ok=True)
 
     # Create writer object for TensorBoard
