+++ intelcamp/algo_main_rnn_v6.py
@@ -2,8 +2,8 @@ import os
 import numpy as np
 import pandas as pd
 import json
-from util import prtime, factors, tile
-import rnn
+from intelcamp.util import prtime, factors, tile
+from intelcamp import rnn
 import seq2seq_model
 
 import torch
@@ -20,7 +20,7 @@ import csv
 import pathlib
 import psutil
 from psutil import virtual_memory
-import buildings_processing as bp
+import intelcamp.buildings_processing as bp
 import logging
 import matplotlib.dates as mdates
 import random
