+++ intelcamp/testing_round_parameters.py
@@ -11,7 +11,7 @@ iterable_type (type) - The type of variable "iterable" as it is defined in confi
 """
 
 import json
-import entry_point as epb
+import intelcamp.entry_point as epb
 import os
 runs = {}
