+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "intelligentcampus"
+name = "intelcamp"
 version = "0.1.0"
 description = ""
 authors = ["Your Name <you@example.com>"]
