+++ tests/fixtures/test_configs.json
@@ -27,8 +27,8 @@
  "train_batch_size": 5,
  "val_batch_size": 1,
  "convert_csvs": false,
- "results_dir": "EnergyForecasting_Results",
- "data_dir": "tests/data",
+ "exp_dir": "",
+ "data_dir": "",
  "resample_freq": 15,
  "sequence_freq_min": 15,
  "splicer": {
