+++ tests/integration_tests/test_training.py
@@ -1,11 +1,12 @@
 import json
 import pathlib
-import entry_point as epb
+import intelcamp.entry_point as epb
 import pytest
-from util import get_exp_dir 
+
 
 TESTS_PATH = pathlib.Path(__file__).parents[1]
 TESTS_FIXTURES_PATH = TESTS_PATH / "fixtures"
+TESTS_DATA_PATH = TESTS_PATH / "data"
 
 
 @pytest.fixture
@@ -15,6 +16,8 @@ def config_for_tests():
     """
     with open(TESTS_FIXTURES_PATH / "test_configs.json", "r") as read_file:
         configs = json.load(read_file)
+
+    configs["data_dir"] = str(TESTS_DATA_PATH)
 
     return configs
 
@@ -47,14 +50,13 @@ def test_model_trains(config_for_tests, tmpdir, config_patch):
     """
     # patch configs and create temporary, unquie output file
     config_for_tests.update(config_patch)
-    config_for_tests["results_dir"] = str(tmpdir / "train_results")
+    exp_dir = pathlib.Path(tmpdir) / "train_results"
+    config_for_tests["exp_dir"] = str(exp_dir)
 
     # train model
     epb.main(config_for_tests)
 
     # check result file were created
-    exp_dir = get_exp_dir(config_for_tests)
-
     assert (exp_dir / "output.out").exists()
     assert (exp_dir / "torch_model").exists()
     assert (exp_dir / "train_stats.json").exists()
