+++ tests/integration_tests/test_validation.py
@@ -2,12 +2,12 @@ import pathlib
 import shutil
 import pytest
 import json
-from util import get_exp_dir 
 
-import entry_point as epb
+import intelcamp.entry_point as epb
 
 TESTS_PATH = pathlib.Path(__file__).parents[1]
 TESTS_FIXTURES_PATH = TESTS_PATH / "fixtures"
+TESTS_DATA_PATH = TESTS_PATH / "data"
 
 @pytest.fixture
 def config_for_tests():
@@ -16,6 +16,8 @@ def config_for_tests():
     """
     with open(TESTS_FIXTURES_PATH / "test_configs.json", "r") as read_file:
         configs = json.load(read_file)
+
+    configs["data_dir"] = str(TESTS_DATA_PATH)
 
     return configs
 
@@ -37,8 +39,8 @@ def test_validation(config_for_tests, tmpdir, config_patch, test_exp_dir):
     config_for_tests.update(config_patch)
     config_for_tests["use_case"] = "validation"
     
-    config_for_tests["results_dir"] = str(tmpdir / "train_results")
-    exp_dir = get_exp_dir(config_for_tests)
+    exp_dir = pathlib.Path(tmpdir) / "train_results"
+    config_for_tests["exp_dir"] = str(exp_dir)
     shutil.copytree(test_exp_dir, exp_dir)
 
     epb.main(config_for_tests)
