+++ tests/unit_tests/test_input_creation.py
@@ -3,11 +3,12 @@ import json
 import pathlib
 import pandas as pd
 
-import entry_point as epb
-from util import get_exp_dir
+import intelcamp.entry_point as epb
+
 
 TESTS_PATH = pathlib.Path(__file__).parents[1]
 TESTS_FIXTURES_PATH = TESTS_PATH / "fixtures"
+TESTS_DATA_PATH = TESTS_PATH / "data"
 
 @pytest.fixture
 def config_for_tests():
@@ -17,16 +18,18 @@ def config_for_tests():
     with open(TESTS_FIXTURES_PATH / "test_configs.json", "r") as read_file:
         configs = json.load(read_file)
 
+    configs["data_dir"] = str(TESTS_DATA_PATH)
+
     return configs
     
     
 def test_create_input_dataframe(config_for_tests, tmpdir):
     # patch configs and create temporary, unquie output file
-    config_for_tests["results_dir"] = tmpdir / "train_results"
+    exp_dir = pathlib.Path(tmpdir) / "train_results"
+    config_for_tests["exp_dir"] = str(exp_dir)
+    exp_dir.mkdir(parents=True, exist_ok=True)
 
     # creat data frame
-    exp_dir = get_exp_dir(config_for_tests)
-    exp_dir.mkdir(parents=True, exist_ok=True)
     train_df, val_df = epb.create_input_dataframe(config_for_tests)
 
     excepted_data_columns = []
