+++ tests/unit_tests/test_util.py
@@ -1,4 +1,4 @@
-import util
+from intelcamp import util
 
 def test_factor():
     assert util.factors(21) == {1, 3, 7, 21}
